requirements.txt
# Dependencias principales
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
cryptography==41.0.3
pycryptodome==3.19.0
tabulate==0.9.0
matplotlib==3.7.1
tqdm==4.65.0
requests==2.31.0

# Utilidades
python-dotenv==1.0.0
gunicorn==21.2.0

# Para la visualización
plotly==5.18.0
numpy==1.24.3

# Visualización y análisis
bcrypt==4.0.1

# API y backend
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5 